requirement.txt
python-telegram-bot==20.7
google-generativeai==0.5.4
